requirements.txt
jupyter
pandas
numpy
scikit-learn==0.21.3
matplotlib==3.1.1
fire==0.2.1
PyYAML==5.1.2